pyproject.toml
[tool.poetry]
name = "multimodal-stockprice-prediction"
version = "0.1.0"
description = ""
authors = ["xavsant <xavierboons2021@economics.smu.edu.sg>"]
license = "BSD 3-Clause License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
numpy = "2.1.0"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
tensorflow = "^2.18.0"
keras = "^3.8.0"
statsmodels = "^0.14.4"
notebook = "^7.3.2"
bs4 = "^0.0.2"
transformers = "^4.49.0"
sentence-transformers = "^3.4.1"
sentencepiece = "^0.2.0"
tf-keras = "^2.19.0"
accelerate = "0.26.0"
wordcloud = "^1.9.4"
yfinance = "^0.2.55"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
